requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
runpod==1.6.2
Pillow==10.4.0
numpy==1.26.4

# PaddleOCR core
paddleocr==2.9.1

# Optional extras commonly needed by PaddleOCR (opencv is used under the hood)
opencv-python-headless==4.10.0.84
scipy==1.11.4
shapely==2.0.5
PyMuPDF==1.24.9
pyclipper==1.3.0.post5
rapidfuzz==3.9.6
python-levenshtein==0.25.1


